+++ graph_examples/triangle_adjacency.txt
@@ -1,3 +1,3 @@
-1 0 2 
+1 2 0 
 2 1 0 
 0 1 2
