+++ graph_examples/triangle_edge_list.txt
@@ -1,3 +1,3 @@
-0 1
 0 2
+0 1
 1 2
