+++ graphtool/graph/__init__.py
@@ -1,0 +1,3 @@
+from .graph import Graph
+from .vertex_edge import Vertex, Edge
+from .orientedGraph import OrientedGraph
