+++ graphtool/graph/_parsing.py
@@ -1,0 +1,7 @@
+import csv
+
+def parse_node_data(filename:str):
+    return
+
+def parse_edge_data(filename:str):
+    return
