+++ graphtool/graph/graph.py
@@ -1,71 +1,6 @@
 from random import random, randint, uniform
-
-
-class Vertex:
-    def __init__(self, name=None, data=None):
-        self.name = name
-        self.data = data
-
-    def __eq__(self, other):
-        return self.name == other.name
-
-    def __str__(self):
-        return str(self.name)
-
-    def __repr__(self):
-        return "V("+str(self.name)+")"
-
-    def __hash__(self):
-        return hash(self.name)
-
-
-class Edge:
-    def __init__(self, *args, **kwargs):
-        """
-        Different ways to initialize an Edge:
-        - Edge(edge)
-        - Edge(iterable of length 2) containing either two Vertex objects or
-            two names of vertices
-        - Edge(a,b) where a and b are either Vertex objects or vertices' names
-        """
-        self.data = kwargs.get("data", None)
-        self.oriented = kwargs.get("oriented", False)
-        self.weight = kwargs.get("weight", 1)
-        if len(args) == 1 and isinstance(args[0], Edge):
-            self.start = args[0].start
-            self.end = args[0].end
-            self.data = args[0].data
-            self.oriented = args[0].oriented
-        else:
-            a, b = None, None
-            if len(args) == 0:
-                a, b = kwargs.get("start", None), kwargs.get("end", None)
-                if a is None or b is None:
-                    raise Exception("Invalid argument")
-            elif len(args) == 1:
-                a, b = args[0][0], args[0][1]
-            elif len(args) == 2:
-                a, b = args[0], args[1]
-            else:
-                raise Exception("Too many arguments : only 2 were expected")
-            if not isinstance(a, Vertex):
-                a = Vertex(a)
-            if not isinstance(b, Vertex):
-                b = Vertex(b)
-            self.start = a
-            self.end = b
-
-    def __eq__(self, other):
-        if self.oriented:
-            return self.stat == other.start and self.end == other.end
-        return {self.start, self.end} == {other.start, other.end}
-
-    def __repr__(self):
-        return "Edge("+str(self.start)+", "+str(self.end)+")"
-
-    def __hash__(self):
-        return hash((self.start, self.end, self.oriented))
-
+from .vertex_edge import Vertex, Edge
+from ._parsing import *
 
 class Graph:
     """
@@ -518,28 +453,3 @@ class Graph:
         v_nb = len(self.vertices())
         possible_edges = v_nb*(v_nb-1)/2
         return e_nb / possible_edges
-
-
-class OrientedGraph:
-    """
-    TODO
-    """
-
-    def __init__(self):
-        pass
-
-    """
-    @staticmethod
-    def erdos_renyi_edge(N, V):
-        adj = [[0 for i in range(N)] for j in range(N)]
-        possible_edges = [(i, j) for j in range(N) for i in range(j)]
-        for i in range(V):
-            (a, b) = possible_edges.pop(randint(0, len(possible_edges) - 1))
-            adj[a][b] = 1
-        return Graph.from_adjacency_matrix(adj)
-
-    def erdos_renyi_proba(N, p):
-        adj = [[1 if uniform(0, 1) < p else 0 for i in range(N)]
-               for j in range(N)]
-        return Graph.from_adjacency_matrix(adj)
-    """
